reafctor home component
