+++ src/consts/paths.ts
@@ -1,0 +1,1 @@
+export const PATHS = [{ path: "/home", label: "Tasks", icon: "mdiHome" }];
